[cairo] refactor text to own module
